Updated spell_check functions to support special chars
Unit tests for both parse_txt() and check_word() were updated to check logic surrounding special characters.

parse_text() was updated to split using a regular expression looking for any character that is not a number, letter or '. It then removes and
spaces or blank entries but keep all special characters in the list.

check_word() was updated to check if a entry in the list is a special character before checking the spelling. If it is a special character, than
entry is skipped.