DataBase Creation Functions
These functions can be used to build the dictionary database for any language.
The DB will be built using instances of Word Class for each word in the dictionary.
These functions will not need to be used during the execution of the full spell check tool
  only during the preparation of the dictionary.
These functions should be language agnostic.