Word Class Refactor
As I was getting into writing Unit Tests for functions using the built dictionary of Word object, I
realized that I wasn't able to laod the save JSON format of the dictionary back into the class objects
properly. It looked like an object but none of the class functions were working.
This refactor adjusts how the class constructor operates in order to allow and expanded dict (**d) to
properly load the Word object.
To go along with this, I wrote a loadDictionary() function which takes in a json file, unpacks it and
then loads the contents into the desired dict of Word object. This function can be used for setting up
Unit Tests as well as be the initial function that runs when our tool starts up to load the dictionaries.