Backeng Integration & Cleanup
To start, this commit touches a lot of the files in the project. Many files just contain
   little changes to fix or cleanup issues that come up during integration.
A new function, logicCntrl(), was added to spell_check.py. This contains all of the logic
   for determining how and where to check for standard spelling errors, context spelling
   errors and getting recommendations. It is a little messy, but recognized as some
   technical debt that will be refactored in the next sprint as we move into the tool
   becoming a 'live' spell checker.
A new file was also created called testset_main.py. This is an executable python script
   that takes in arguments for the JSON DictDatabase, a test set (formatted like the
   ones provided by Dr. Scannell, and a language specifier. The logic within this script
   mirrors the logic in logicCntrl() but handles data in the form of a test set rather
   than a raw user input. I also want to both clean this file up going forward and add
   some metrics calculations to better asses improvements to the spell checker.
There is also an updated version of the English Dictionary in this commit. A small issue
   was found in the parser since the last database was built and so this updates corrects
   those errors.
The rest of the changes are small and should be relatively self explanitory.