Final release (#76)
* HTML Cleanup

* At a distance checks

* Footer and fixes

* Adding check_other_lang tests (#52)

Added tests for check_other_lang based on the current dictionary

* Modifying Distance Recs (#53)

Modified distance recs to work with Irish as well. When building the
languagehelper object pass the desired dictionary for the langauge along
with the langauge itself (at the moment, irish or english all lowercase)

Can then call getSuggestions for edit distance 1 or getSuggestionsExtra
for edit distance 2

* Backeng Integration & Cleanup

To start, this commit touches a lot of the files in the project. Many files just contain
   little changes to fix or cleanup issues that come up during integration.
A new function, logicCntrl(), was added to spell_check.py. This contains all of the logic
   for determining how and where to check for standard spelling errors, context spelling
   errors and getting recommendations. It is a little messy, but recognized as some
   technical debt that will be refactored in the next sprint as we move into the tool
   becoming a 'live' spell checker.
A new file was also created called testset_main.py. This is an executable python script
   that takes in arguments for the JSON DictDatabase, a test set (formatted like the
   ones provided by Dr. Scannell, and a language specifier. The logic within this script
   mirrors the logic in logicCntrl() but handles data in the form of a test set rather
   than a raw user input. I also want to both clean this file up going forward and add
   some metrics calculations to better asses improvements to the spell checker.
There is also an updated version of the English Dictionary in this commit. A small issue
   was found in the parser since the last database was built and so this updates corrects
   those errors.
The rest of the changes are small and should be relatively self explanitory.

* PR Comments Cleanup

* Building scraper and dictionary (#63)

Built a web scraper that scrapes all unique valid words from the eDIL
Irish dictionary, also reads a pdf titled gv2ga which is an additional
dictionary as eDIL is primarily a dictionary of the Irish language prior
to the year 1700 and gv2ga should be more current

* Changing color of context underline (#64)

Changed the color of underlining used for words determined to be wrong
in context

* Added Metrics Calculations & Expanded Dict Building

Added the ability to read in a file with answers in testset_main.py and to calcluate metrics of the results of
  spell checking.
Added ability to initialize a dictionary database with only words and no context.
Added option (currently commented out) to assemble database based on initialized db where word array only adds
  context and not new words to the dictionary. Currently this does not improve effectiveness but leaving in so
  that it can be revisted later during testing.

* Fixed broken unit test

* Updating English Dictionary Database

* Dual language feature

* updated requirements.txt

* Altering lang_select method

Altered each instance of lang_select being called to pass the clicked
HTML tag to the lang_select method so the selected language can be shown
properly translated. Also altered the lang_select method to properly
handle the new data being passed in

* translation dropdown

* Finalizing page lang change

Finalized the ability to change the language of the page by creating a
process_lang route in views.py which sets the session variable to the
value passed to it by the ajax request in the new site_lang function
present in base.html which then reloads the page forcing get_locale to
check if a session variable named language is avaiable and if so sets
the locale to that language

* Fixed issue

Was not working on index.html, found the issue, leftover line from
earlier testing was clearing all session variables when the page loaded.
Have removed that line and some unneeded commented code

* Removing unneeded delay

Removed an unneeded delay from base.html, should save roughly 200
milliseconds per reload

* Created trieNode Class

This class is for building a trie structure to hold all words in the corpus.
Along with the class there is a function to build the trie based on a given list of words,
  and a function to return all words in trie the are a specific edit distance away from a
  given word.

* Implemented Partial Use of Trie in testset_main.py

Updated logic within testset_main.py to bring in the Error Model triples and use when processing Irish language.
Still Neede: Currently only implemented for Irish, have not yet worked on it for English.
             When an incorrect word is identified, if it is not in the Triple, then we don't currently try
               to give a suggestion. Needs to be implemented using Trie Edit Distance and some form of sorting.

* Fix Import

* Corrected Some Logic in testset_main.py

* Error Model Class, Unit Tests, and EM Creation Executable

* logicCntrl() Refactor and Simplfied check_word()

* testset_main.py Refactor

* Copy of Irish DictDB That has been Performing Best

We can clean up exactly which dictionaries and error_models before the final release

* Updated views.py to handle new logicCntrl() function

Made update to logicCntrl to return whether the spelling mistake was due to conext error or a true misspelled word.
Made update to error model so that scenarios where only a single recomendation is present, it does not show up twice on the rec list

* English Error Model file

* Fixed broken Error Model Unit Tests

* Bug Fix (#70)

Fixed a bug created by switching page langauge

* Word Replacement Upon Submit

* Fixed bug in testset_main executable

* Fixed issue with context sensitive checks

* Constructing data passage

Built ability to get data passed to wordRecs.js from index.html.
Currently it will generate a list but only for the first misspelled
word, and it is generated in the wrong position

* Updating data passage

Updating data passage so that multiple words are able to be clicked on
and a list is generated

* Word Replacing

Fixed word replacement functionality, have discovered a bug where words
may or may not be able to properly access the recommends list based on
where it is in the string, need to try and figure out how to fix

* fixed dropdown suggestions

* button fix

* moved button into form tag

* Small tweaks to dictionary building script and cntrlLogic()

* Update Dictionary and Error Model Files

These are the final files after tuning for the best results

* Updating loaded json

* Updating tests to match new Dictionary

* Fix Errors and add enter submit

Fixed errors with file names and added submission on enter key

* Adding lost hot fix

adding lost hot fixe

Co-authored-by: bmacgavin <71277163+bmacgavin@users.noreply.github.com>
Co-authored-by: Andrew Beiler <andrewbeiler@Andrews-Air.attlocal.net>
Co-authored-by: Drew Beiler <89706237+abeiler7@users.noreply.github.com>
Co-authored-by: Andrew Beiler <andrewbeiler@Andrews-MacBook-Air.local>